feat: use try catch instead for cleaner code
