+++ apps/service-scrape/function-batch_scrape_suburbs/src/service/Domain/DomainSuburbService.ts
@@ -80,28 +80,25 @@ export class DomainSuburbService extends IService {
      * @returns object containing rawSuburbSchema
      */
     tryExtractProfile(args: { nextDataJson: object }) {
-        return this.tryCatchLogException(
-            args,
-            () => {
-                const validNextjson = nextDataJsonSchema.parse(args.nextDataJson)
-                const intestingObjects = Object.values(
-                    validNextjson.props.pageProps.__APOLLO_STATE__,
-                )
-                return {
-                    ..._rawSuburbSchema.parse({
-                        suburb: intestingObjects[0],
-                        location: intestingObjects[1],
-                    }),
-                    boundaryGeoJson: _boundaryGeoJsonSchema.parse(
-                        JSON.parse(
-                            intestingObjects[0]['suburbShape({"geometryPrecision":"high"})']
-                                .boundaryGeoJson,
-                        ),
+        try {
+            const validNextjson = nextDataJsonSchema.parse(args.nextDataJson)
+            const intestingObjects = Object.values(validNextjson.props.pageProps.__APOLLO_STATE__)
+            return {
+                ..._rawSuburbSchema.parse({
+                    suburb: intestingObjects[0],
+                    location: intestingObjects[1],
+                }),
+                boundaryGeoJson: _boundaryGeoJsonSchema.parse(
+                    JSON.parse(
+                        intestingObjects[0]['suburbShape({"geometryPrecision":"high"})']
+                            .boundaryGeoJson,
                     ),
-                } satisfies DomainListingsDTO
-            },
-            null,
-        )
+                ),
+            } satisfies DomainListingsDTO
+        } catch (e) {
+            this.logException('error', e, args)
+            return null
+        }
     }
 
     /**
@@ -110,37 +107,36 @@ export class DomainSuburbService extends IService {
      * @returns Object containing tables for database inserts
      */
     tryTransformProfile(args: { rawSuburbData: DomainListingsDTO }) {
-        return this.tryCatchLogException(
-            args,
-            () => {
-                // Simplify geo boundary to reduce storage size.
-                const compressedCoordinates = simplify(
-                    polygon(args.rawSuburbData.boundaryGeoJson.coordinates),
-                    {
-                        tolerance: 0.00025,
-                        highQuality: true,
-                    },
-                ).geometry.coordinates
+        try {
+            // Simplify geo boundary to reduce storage size.
+            const compressedCoordinates = simplify(
+                polygon(args.rawSuburbData.boundaryGeoJson.coordinates),
+                {
+                    tolerance: 0.00025,
+                    highQuality: true,
+                },
+            ).geometry.coordinates
 
-                const boundaryCoord = createPostgisPolygonString(compressedCoordinates[0])
-                if (!boundaryCoord) throw new Error('invalid locality boundary coordinate')
+            const boundaryCoord = createPostgisPolygonString(compressedCoordinates[0])
+            if (!boundaryCoord) throw new Error('invalid locality boundary coordinate')
 
-                // rawSuburbData.suburb.statistics.population
-                // rawSuburbData.suburb.statistics.ownerOccupierPercentage
-                // rawSuburbData.suburb.statistics.marriedPercentage
+            // rawSuburbData.suburb.statistics.population
+            // rawSuburbData.suburb.statistics.ownerOccupierPercentage
+            // rawSuburbData.suburb.statistics.marriedPercentage
 
-                // rawSuburbData.location.data.propertyCategories
-                // rawSuburbData.suburb.schools
-                return {
-                    localities_table: {
-                        postcode: args.rawSuburbData.suburb.postcode,
-                        suburb_name: args.rawSuburbData.suburb.name,
-                        state_abbreviation: args.rawSuburbData.suburb.state,
-                        boundary_coordinates: boundaryCoord,
-                    } satisfies Updateable<Schema.LocalitiesTable>,
-                }
-            },
-            null,
-        )
+            // rawSuburbData.location.data.propertyCategories
+            // rawSuburbData.suburb.schools
+            return {
+                localities_table: {
+                    postcode: args.rawSuburbData.suburb.postcode,
+                    suburb_name: args.rawSuburbData.suburb.name,
+                    state_abbreviation: args.rawSuburbData.suburb.state,
+                    boundary_coordinates: boundaryCoord,
+                } satisfies Updateable<Schema.LocalitiesTable>,
+            }
+        } catch (e) {
+            this.logException('error', e, args)
+            return null
+        }
     }
 }
