--- apps/service-scrape/function-batch_scrape_suburbs/src/service/scrapeUtil.ts
@@ -1,52 +1,0 @@
-import { load } from 'cheerio'
-import { traceTryFunction } from '../instrumentation'
-import z from 'zod'
-
-export const scrapeUtil = {
-    /**
-     * @description Extract nicely formatted hydration JSON from <script id="__NEXT_DATA__">
-     * @param html
-     */
-    tryExtractNextJson(html: string) {
-        return traceTryFunction('scrapeUtil.tryExtractNextJson', arguments, 'ERROR', async () => {
-            const $ = load(html)
-            const nextJson = $('script[id="__NEXT_DATA__"]').text()
-            if (nextJson === '') throw Error('Cannot extract Next.js JSON')
-            return JSON.parse(nextJson) as object
-        })
-    },
-
-    /**
-     * Parse raw datetime to ISO format
-     * @param datetime
-     * @returns
-     */
-    parseDatetime: (datetime: string | null | undefined) => {
-        if (datetime) {
-            return z.iso.datetime().parse(`${datetime}Z`)
-        } else {
-            return null
-        }
-    },
-
-    /**
-     * @returns ISO standard date
-     */
-    currentDatetime: () => new Date().toISOString(),
-
-    /**
-     * @description Homes tend to sell at higher price.
-     * @returns Integer price
-     */
-    highestPriceFromString(priceString: string) {
-        const prices = priceString
-            .replaceAll(/[^0-9^ ^-^$^.]+/g, '') // Expect numbers separated by '_' or ' '
-            .matchAll(/[$]( )?\d+/g) // Integer price starts with $, could have a space
-            .toArray()
-        if (prices.length === 0) return null
-        const priceList = prices.map((match) =>
-            Number.parseFloat(match.toString().replaceAll('$', '')),
-        )
-        return Math.max(...priceList)
-    },
-}
